+++ task4/print.cpp
@@ -11,13 +11,13 @@ int main(int argc, char ** argv) {
         return -1;
     }
     fstream file(argv[1], ios::binary | ios::in);
-    uint64_t n, m;
+    int n, m;
     file.read((char *)&n, sizeof(n));
     file.read((char *)&m, sizeof(m));
     cout << setw(5) << n << setw(5) << m << endl;
     double f;
-    for (uint64_t i = 0; i < n; i++) {
-        for (uint64_t j = 0; j < m; j++) {
+    for (int i = 0; i < n; i++) {
+        for (int j = 0; j < m; j++) {
             file.read((char *)&f, sizeof(f));
             cout << fixed << setw(15) << setprecision(3) << f;
         }
